Move CFA generation to separate files
